+++ .gitignore
@@ -86,4 +86,4 @@ tests/test-sampling
 tests/test-tokenizer-0-llama
 tests/test-tokenizer-0-falcon
 tests/test-tokenizer-1-llama
-tests/test-tokenizer-1-falcon
+tests/test-tokenizer-1-bpe
